refactor(jwt-local.constants.ts): centralizes file name
